requirements.txt
hypothesis # seems to be required for numpy, without it the tests fail
numpy
pypng
svgwrite
tqdm
scikit-image
tensorflow-macos
tqdm
matplotlib
